Documented `bucket` option
Change log:
1. Added documentation for `-mode bucket`

Summary:
`bucket` mode can now be used to launch pipeline on specific folder located outside of Ardetype and for it to be saved in path,
defined in `config_modular.yaml/output_bucket_path`.
The option is not `idempotent` - if the processing fails, the output stays in `Ardetype` folder to be manually checked.